# storage/api-docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AI4BMI",
        "description": "API E-commerce pour l'app mobile (clients). Authentification : Bearer token (Sanctum) ; seuls POST /register et POST /login sont publics.\n\n**Cycle commande** : création (POST /orders) → statut `pending` → paiement confirmé (webhook ou back-office) → `paid` → expédition → `shipped`. Annulation possible uniquement si `pending` (POST /orders/{id}/cancel). Adresse de livraison : fournie à la commande ou reprise du profil (PATCH /user).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Liste des produits",
                "description": "Catalogue. Filtres : category_id, search, per_page. Authentification requise.",
                "operationId": "1f38848511def488712c7ce0b68d389c",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "E-commerce"
                ],
                "summary": "Détail d'un produit",
                "description": "Authentification requise.",
                "operationId": "d34acc37dcebc067282733da735f4c4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/webhooks/payment": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Confirmation de paiement (webhook)",
                "description": "Appelé par le prestataire de paiement ou le back-office pour passer une commande en « paid ». Secret requis (PAYMENT_WEBHOOK_SECRET). En production, préférer une signature HMAC.",
                "operationId": "43a358f197e6ff0e8590e8113d2d1508",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "secret"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "ID de la commande",
                                        "type": "integer"
                                    },
                                    "secret": {
                                        "description": "Clé secrète (PAYMENT_WEBHOOK_SECRET)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commande marquée comme payée"
                    },
                    "400": {
                        "description": "Secret invalide ou manquant"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    },
                    "422": {
                        "description": "Commande non en statut pending"
                    }
                }
            }
        },
        "/webhooks/fedapay": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook FedaPay",
                "description": "Appelé par FedaPay lors d'événements (transaction.approved). Passe la commande en « paid ». Configurer l'URL dans le dashboard FedaPay (compte test ou live).",
                "operationId": "1b270b6ca5298b2b2fb11ccf1dcc7bfd",
                "requestBody": {
                    "description": "Payload FedaPay (event, data.transaction...)"
                },
                "responses": {
                    "200": {
                        "description": "Traité"
                    },
                    "400": {
                        "description": "Payload invalide"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}